+++ lib/whisper_ggml.dart
@@ -1,5 +1,4 @@
+export 'src/models/_models.dart';
 export 'src/models/whisper_model.dart';
 export 'src/whisper.dart';
-export 'src/whisper_audio_convert.dart';
 export 'src/whisper_controller.dart';
-export 'src/models/_models.dart';